20200720 22:31 lu

+++ qy109-project-cehui-management/qy109-project-cehui-mapper/src/main/java/com/aaa/cehui/mapper/RoleMenuMapper.java
@@ -11,7 +11,7 @@ import tk.mybatis.mapper.common.Mapper;
  **/
 public interface RoleMenuMapper extends Mapper<RoleMenu> {
 
-    Integer deleteMenuByRoleId(Integer roleId);
+    Integer deleteMenuByRoleId(Long roleId);
 
-    Integer insertMenuByRoleId(Integer roleId,Integer menuId);
+    Integer insertMenuByRoleId(Long roleId,Integer menuId);
 }
